requirements.txt
pandas
numpy
streamlit
matplotlib
scikit-learn
xgboost
tensorflow
statsmodels
newsapi-python
yfinance